requirements.txt
# filepath: /Users/amac/Documents/code/Labeeb/requirements.txt

# Labeeb requirements are now managed in setup.py
# Please install using:
#   pip install .
# For development tools:
#   pip install .[dev]
# For testing tools:
#   pip install .[test]

# Core dependencies
aiohttp>=3.12.0
pyautogui>=0.9.54
psutil>=7.0.0
pillow>=11.2.1
pytesseract>=0.3.10  # For OCR
opencv-python>=4.8.0  # For image processing
numpy>=1.24.0
websockets>=12.0  # For WebSocket support
ollama>=0.1.7
transformers>=4.40.0
pyjwt>=2.8.0
bcrypt>=4.1.2
orjson>=3.9.10
requests>=2.31.0
urllib3>=2.0.0
cryptography>=41.0.0
toml>=0.10.2

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1

# Note: grpcio and pymilvus removed as they're not needed for native tools

# GUI dependencies
# tkinter  # Usually comes with Python

# Testing dependencies
pytest-asyncio>=0.18.0

# Development dependencies
python-json-logger>=2.0.7
colorlog>=6.7.0

# Awareness Module Dependencies
screen-brightness-control>=0.16.0
pywin32>=306; platform_system == "Windows"
python-xlib>=0.33; platform_system == "Linux"
pyobjc-framework-Quartz>=9.2; platform_system == "Darwin"

# Labeeb requirements (Ubuntu 24.04+)
# GUI and audio dependencies
PyQt5>=5.15.11
PyQt5-sip>=12.17.0
PyQt5-Qt5>=5.15.17
pyaudio>=0.2.13
openai-whisper @ git+https://github.com/openai/whisper.git
# For screenshots and image processing
Pillow>=9.0.0
# Add other dependencies below
pytest==8.0.0
pytest-asyncio==0.23.5

# Audio and Speech
pyttsx3>=2.90
whisper>=1.1.10

# Optional Dependencies
# For macOS: say (built-in)
# For Linux: espeak (system package)
# For Windows: SAPI.SpVoice (built-in)

PyQt5>=5.15.0
ollama>=0.1.0
transformers>=4.30.0
aiohttp>=3.8.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.0.0
colorama>=0.4.6

# Web and browser automation
selenium>=4.15.0
playwright>=1.40.0
bs4>=0.0.1
beautifulsoup4>=4.12.0

# System and hardware interaction
pynput>=1.7.6
keyboard>=0.13.5
pygetwindow>=0.0.9
pyudev>=23.2.1
usb>=2.0.0
sounddevice>=0.4.6

# Data processing and visualization
matplotlib>=3.8.0
networkx>=3.2.0
torch>=2.1.0
spacy>=3.7.0
langdetect>=1.0.9
distro>=1.9.0

# Platform specific
# macOS
pyobjc-framework-AVFoundation>=9.2
pyobjc-framework-CoreAudio>=9.2
pyobjc-framework-Foundation>=9.2
pyobjc-framework-IOBluetooth>=9.2
pyobjc-framework-IOKit>=9.2

# Windows
pywin32>=306; platform_system == "Windows"
wmi>=1.5.1; platform_system == "Windows"

# FFmpeg
ffmpeg-python>=0.2.0

# Agent frameworks
smolagents>=0.1.0
